### 2024-08-05 #7072 @editorialbot
[PRE REVIEW]: ggoncoplot: an R package for interactive visualisation of somatic mutation data from cancer patient cohorts
**Submitting author:** @selkamand (<a href="http://orcid.org/0000-0003-2270-8088">Sam El-Kamand</a>)
**Repository:** https://github.com/selkamand/ggoncoplot
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
