### 2024-08-07 selkamand/ggoncoplot#140 @selkamand
Add CONTRIBUTING.md
Move contribution guidelines from the README.RMD to CONTRIBUTING.md
See https://github.com/openjournals/joss-reviews/issues/7072 for details
