### 2024-08-07 selkamand/ggoncoplot#141 @selkamand
JOSS: fix missing DOIs
See pre-review
https://github.com/openjournals/joss-reviews/issues/7072
