### 2024-08-07 c392b31
docs: moved community contribution guidelines to CONTRIBUTING.md
