### 2024-08-07 #7072 @selkamand
Thanks @Kevin-Mattheus-Moerman. Missing DOIs and a CONTRIBUTING.md have been added
