### 2024-08-07 #7072 @Kevin-Mattheus-Moerman
@selkamand thanks for making those changes. I've just added the `query-scope` label which will trigger a scope review by the editorial board. The objective of the scope review is to determine if the functionality offered by the work is substantial enough, is research software, and conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This is normal for projects that are on the small side and also for projects that appear to focus on visualization. The scope review should take about 2 weeks to complete.
