### 2024-08-19 #7072 @Kevin-Mattheus-Moerman
@selkamand I am pleased to inform the board has concluded this work may be in scope for JOSS. We will now proceed to seek a handling editor.
