### 2024-09-05 #7072 @ppxasjsm
Hi @@duncanjunegong would you be willing to review this JOSS submission?
