### 2024-09-12 #7072 @lpantano
Hi, I am editor for joss, and my colleague @zainabarshad, can be a reviewer for this tool.
