### 2024-09-27 13c9b5c
docs: fixed link to comparison pdf
