+++ README.Rmd
@@ -139,7 +139,7 @@ Oncoplots are highly effective for visualising mutation data in cancer cohorts b
 We developed ggoncoplot as the first R package to address all these challenges together (\autoref{fig:comparison}). Examples of all key features are available in the [ggoncoplot manual](https://selkamand.github.io/ggoncoplot/articles/manual.html).
 
 
-A full comparison of ggoncoplot features with similar tools is available [here](papre/ggoncoplot_comparison.pdf)
+A full comparison of ggoncoplot features with similar tools is available [here](paper/ggoncoplot_comparison.pdf)
 
 
 
