+++ pkgdown.yml
@@ -1,10 +1,10 @@
 pandoc: 3.1.11
-pkgdown: 2.1.0
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   manual: manual.html
   sorting_algorithm: sorting_algorithm.html
-last_built: 2024-08-07T06:43Z
+last_built: 2024-09-27T00:40Z
 urls:
   reference: https://selkamand.github.io/ggoncoplot/reference
   article: https://selkamand.github.io/ggoncoplot/articles
