### 2024-10-22 #7072 @ppxasjsm
Hi @allaway, is there any chance you are available to review this submission for JOSS?
