### 2024-10-22 #7390 @editorialbot
[REVIEW]: ggoncoplot: an R package for interactive visualisation of somatic mutation data from cancer patient cohorts
**Submitting author:** @selkamand (<a href="http://orcid.org/0000-0003-2270-8088">Sam El-Kamand</a>)
**Repository:** https://github.com/selkamand/ggoncoplot
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
**Reviewers:** @zainabarshad, @tushardave26, @welch16, @majensen, @vinegang
