### 2024-11-18 selkamand/ggoncoplot#146 @selkamand
Update `citation()` to JOSS paper
Currently waiting on JOSS review:
https://github.com/openjournals/joss-reviews/issues/7390

Once complete need to update description so citation("ggoncoplot") points to the JOSS paper
