+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   manual: manual.html
   sorting_algorithm: sorting_algorithm.html
-last_built: 2024-09-27T00:40Z
+last_built: 2024-11-20T01:59Z
 urls:
   reference: https://selkamand.github.io/ggoncoplot/reference
   article: https://selkamand.github.io/ggoncoplot/articles
