### 2024-11-25 5634083
fix: updated gg1d call to work with new gg1d
and patchwork (axis alignment now done without label alignment)
