+++ R/oncoplot.R
@@ -10,6 +10,9 @@ utils::globalVariables(
 # Oncoplot ----------------------------------------------------------------
 
 #' ggoncoplot
+#'
+#' Create an oncoplot.
+#'
 #' @importFrom patchwork plot_layout
 #'
 #' @param data data for oncoplot. A data.frame with 1 row per mutation in your cohort. Must contain columns describing gene_symbols and sample_identifiers (data.frame)
@@ -399,46 +402,43 @@ ggoncoplot <- function(data,
       cols_to_plot = cols_to_plot_metadata,
       interactive = FALSE,
       verbose = if(verbose) 1 else 0,
-      cli_header = "Plotting Sample Metadata",
-
-      # Tile width
-      width = options$tile_width,
-
-      # Axis Text Fontsizes
-      fontsize_y_text = options$fontsize_metadata_text,
-
-      # Legend Fontsizes
-      legend_title_size = options$fontsize_metadata_legend_title,
-      legend_text_size = options$fontsize_metadata_legend_text,
-      fontsize_barplot_y_numbers = options$fontsize_metadata_barplot_y_numbers,
-
-      # Legend Layout
-      show_legend_titles = options$show_legend_titles, #default TRUE
-      legend_nrow = options$metadata_legend_nrow, # default NULL
-      legend_ncol = options$metadata_legend_ncol, # default NULL
-      legend_key_size = options$metadata_legend_key_size,
-
-      # Dealing with NAs
-      na_marker = options$metadata_na_marker,
-      na_marker_size = options$metadata_na_marker_size,
-
-      # Processing Levels
+      palettes = metadata_palette,
       maxlevels = options$metadata_maxlevels,
 
-      # Numeric Values
-      numeric_plot_type = options$metadata_numeric_plot_type,
-      legend_orientation_heatmap = options$metadata_legend_orientation_heatmap,
+      options = gg1d::gg1d_options(
+        cli_header = "Plotting Sample Metadata",
 
-      # Palettes
-      palettes = metadata_palette,
-      # colours_default = metadata_colours_default,
-      # colours_default_logical = metadata_colours_default_logical,
-      # colours_missing = metadata_colours_missing,
-      y_axis_position = "left"
+        # Tile width
+        width = options$tile_width,
+
+        # Axis Title Fontsizes
+        fontsize_y_title = options$fontsize_metadata_text,
+
+        # Legend Fontsizes
+        legend_title_size = options$fontsize_metadata_legend_title,
+        legend_text_size = options$fontsize_metadata_legend_text,
+        fontsize_barplot_y_numbers = options$fontsize_metadata_barplot_y_numbers,
+
+        # Legend Layout
+        show_legend_titles = options$show_legend_titles, #default TRUE
+        legend_nrow = options$metadata_legend_nrow, # default NULL
+        legend_ncol = options$metadata_legend_ncol, # default NULL
+        legend_key_size = options$metadata_legend_key_size,
+
+        # Dealing with NAs
+        na_marker = options$metadata_na_marker,
+        na_marker_size = options$metadata_na_marker_size,
+
+        # Numeric Values
+        numeric_plot_type = options$metadata_numeric_plot_type,
+        legend_orientation_heatmap = options$metadata_legend_orientation_heatmap,
+        y_axis_position = "left"
+      )
     )
[…]
