+++ DESCRIPTION
@@ -19,12 +19,12 @@ Imports:
     cli,
     dplyr,
     forcats,
-    gg1d (>= 0.0.0.9000),
+    gg1d (>= 0.1.0),
     ggiraph,
     ggplot2,
     grid,
     mutationtypes (>= 0.0.0.9000),
-    patchwork,
+    patchwork (>= 1.3.0),
     RColorBrewer,
     rlang,
     scales
@@ -43,7 +43,6 @@ VignetteBuilder: knitr
 Config/testthat/edition: 3
 Remotes: 
     selkamand/mutationtypes,
-    selkamand/gg1d,
     selkamand/express,
     CCICB/somaticflags
 Roxygen: list(markdown = TRUE)
