### 2024-12-13 a42173c
fix: dropped redundant remotes
mutationtypes and somaticflags are both available from CRAN
