+++ DESCRIPTION
@@ -23,7 +23,7 @@ Imports:
     ggiraph,
     ggplot2,
     grid,
-    mutationtypes (>= 0.0.0.9000),
+    mutationtypes,
     patchwork (>= 1.3.0),
     RColorBrewer,
     rlang,
@@ -36,13 +36,11 @@ Suggests:
     express,
     knitr,
     rmarkdown,
-    somaticflags (>= 0.0.0.9000),
+    somaticflags,
     testthat (>= 3.0.0),
     vdiffr
 VignetteBuilder: knitr
 Config/testthat/edition: 3
 Remotes: 
-    selkamand/mutationtypes,
-    selkamand/express,
-    CCICB/somaticflags
+    selkamand/express
 Roxygen: list(markdown = TRUE)
