### 2025-03-05 #7390 @ppxasjsm
Hi @tgerke, I am trying very hard to find another reviewer for this JOSS submission. Is there any chance this fits into your area of remit and you are available to review it? For more details on how to review for JOSS, see the reviewers' guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
