### 2025-03-12 129c584
fix: space left for tmb plot even when not drawn
resolves issue #137
