+++ R/oncoplot.R
@@ -445,7 +445,7 @@ ggoncoplot <- function(data,
     gg_tmb = gg_tmb_barplot,
     gg_gene = gg_gene_barplot,
     gg_metadata = gg_metadata,
-    gg_tmb_height = options$plotsize_tmb_rel_height,
+    gg_tmb_height = if(draw_tmb_barplot) options$plotsize_tmb_rel_height else 0,
     gg_gene_width = options$plotsize_gene_rel_width,
     gg_metadata_height = options$plotsize_metadata_rel_height
     )
