### 2025-03-12 bd4c86e
fix: dropped anonymous functions
Can now use ggoncoplot with R < 4.1

Resolves issue #154
