### 2025-03-12 397e801
docs: improved hexlogo
resolves issue #158
