### 2025-03-12 7dd5120
docs: improved ggoncoplot description
Resolves issue #151
