### 2025-03-12 b082445
fix: space left for margin plots only when drawn
resolves #164
