### 2025-03-13 5a7e3ff
feat: metadata can be visualised above oncoplot
Resolves #163
