+++ R/oncoplot.R
@@ -449,7 +449,8 @@ ggoncoplot <- function(data,
     gg_metadata = gg_metadata,
     gg_tmb_height = if(draw_tmb_barplot) options$plotsize_tmb_rel_height else 0,
     gg_gene_width = if(draw_gene_barplot) options$plotsize_gene_rel_width else 0,
-    gg_metadata_height = if(!is.null(metadata)) options$plotsize_metadata_rel_height else 0
+    gg_metadata_height = if(!is.null(metadata)) options$plotsize_metadata_rel_height else 0,
+    metadata_position = options$metadata_position
     )
 
   ## Control Look of oncoplot + marginal plots
@@ -1116,26 +1117,33 @@ ggoncoplot_tmb_barplot <- function(data, col_samples, col_mutation_type, palette
 #' @param gg_tmb_height percentage of plot height taken up by TMB plot (should be between 5-95) (number)
 #' @param gg_gene_width percentage of plot width taken up by genebar plot (should be between 5-95) (number)
 #' @param gg_metadata_height percentage of plot height taken up by metadata plot (should be between 5-95) (number)
+#' @param metadata_position should metadata plot be on the 'top' or the 'bottom' of the oncoplot?
 #' @return patchwork object (or ggplot obj if both `gg_tmb` and `gg_gene` are NULL)
 #'
-combine_plots <- function(gg_main, gg_tmb = NULL, gg_gene = NULL, gg_metadata = NULL, gg_tmb_height, gg_gene_width, gg_metadata_height){
+combine_plots <- function(gg_main, gg_tmb = NULL, gg_gene = NULL, gg_metadata = NULL, gg_tmb_height, gg_gene_width, gg_metadata_height, metadata_position){
+
   assertions::assert(gg_tmb_height + gg_metadata_height < 95)
   assertions::assert(gg_gene_width < 95)
+  assertions::assert_one_of(metadata_position, c("top", "bottom"));
+
+  # Define top and bottom heights and plots depending on metadata_position
+  gg_top_height <- if(metadata_position == "top") gg_metadata_height else gg_tmb_height
+  gg_bottom_height <- if(metadata_position == "top") gg_tmb_height else gg_metadata_height
+  gg_top_plot <- if(metadata_position == "top") gg_metadata else gg_tmb
+  gg_bottom_plot <- if(metadata_position == "top") gg_tmb else gg_metadata
 
   gg_main_height = 100 - gg_tmb_height - gg_metadata_height
-  gg_main_top = gg_tmb_height + 1
-  gg_main_bottom = gg_tmb_height + 1 + gg_main_height
+  gg_main_top = gg_top_height + 1
+  gg_main_bottom = gg_top_height + 1 + gg_main_height
 
   gg_main_width = 100 - gg_gene_width
-
-
 
   # Define layouts (will need to edit to make layout respect gg_main_height, gg_main_width and gg_metadata_height)
   layout <- c(
     patchwork::area(t = gg_main_top, l = 0, b = gg_main_bottom, r = gg_main_width), # Main Plot
-    if(!is.null(gg_tmb)) patchwork::area(t = 0, l = 0, b = gg_tmb_height, r = gg_main_width) else patchwork::area(), # TMB Barplot
+    if(!is.null(gg_top_plot)) patchwork::area(t = 0, l = 0, b = gg_top_height, r = gg_main_width) else patchwork::area(), # TMB Barplot (or metadata if metadata_position="top")
[…]
