+++ man/ggoncoplot_options.Rd
@@ -50,6 +50,7 @@ ggoncoplot_options(
   pathway_text_angle = 0,
   ggoncoplot_guide_ncol = 2,
   legend_key_size = 0.4,
+  metadata_position = c("bottom", "top"),
   fontsize_metadata_text = 12,
   fontsize_metadata_legend_title = fontsize_legend_title,
   fontsize_metadata_legend_text = fontsize_legend_text,
@@ -191,6 +192,8 @@ only have an effect if \code{genebar_scale_breaks = ggplot2::waiver()}. Use \cod
 
 \item{metadata_legend_orientation_heatmap}{the orientation of heatmaps in legends. One of "horizontal" or "vertical"
 number of breaks given by the transformation.}
+
+\item{metadata_plot_position}{should the metadata plot be at the \strong{top} or \strong{bottom} of the oncoplot.}
 }
 \value{
 ggoncoplot options object ready to be passed to \code{\link[=ggoncoplot]{ggoncoplot()}} \code{options} argument
