### 2025-03-13 f787e40
feat: sampleId angles & position now controllable
Resolves issues #163 and #166
