### 2025-03-13 6b105fc
feat: can buffer tmb/metadata from main plot
resolves #168
