+++ R/oncoplot.R
@@ -452,7 +452,9 @@ ggoncoplot <- function(data,
     gg_tmb_height = if(draw_tmb_barplot) options$plotsize_tmb_rel_height else 0,
     gg_gene_width = if(draw_gene_barplot) options$plotsize_gene_rel_width else 0,
     gg_metadata_height = if(!is.null(metadata)) options$plotsize_metadata_rel_height else 0,
-    metadata_position = options$metadata_position
+    metadata_position = options$metadata_position,
+    buffer_tmb = options$buffer_tmb,
+    buffer_metadata = options$buffer_metadata
     )
 
   ## Control Look of oncoplot + marginal plots
@@ -1126,32 +1128,42 @@ ggoncoplot_tmb_barplot <- function(data, col_samples, col_mutation_type, palette
 #' @param gg_gene_width percentage of plot width taken up by genebar plot (should be between 5-95) (number)
 #' @param gg_metadata_height percentage of plot height taken up by metadata plot (should be between 5-95) (number)
 #' @param metadata_position should metadata plot be on the 'top' or the 'bottom' of the oncoplot?
+#' @inheritParams ggoncoplot_options
 #' @return patchwork object (or ggplot obj if both `gg_tmb` and `gg_gene` are NULL)
 #'
-combine_plots <- function(gg_main, gg_tmb = NULL, gg_gene = NULL, gg_metadata = NULL, gg_tmb_height, gg_gene_width, gg_metadata_height, metadata_position){
+combine_plots <- function(gg_main, gg_tmb = NULL, gg_gene = NULL, gg_metadata = NULL, gg_tmb_height, gg_gene_width, gg_metadata_height, metadata_position, buffer_metadata, buffer_tmb){
 
   assertions::assert(gg_tmb_height + gg_metadata_height < 95)
   assertions::assert(gg_gene_width < 95)
   assertions::assert_one_of(metadata_position, c("top", "bottom"));
 
+  metadata_on_top = metadata_position == "top"
+
   # Define top and bottom heights and plots depending on metadata_position
-  gg_top_height <- if(metadata_position == "top") gg_metadata_height else gg_tmb_height
-  gg_bottom_height <- if(metadata_position == "top") gg_tmb_height else gg_metadata_height
-  gg_top_plot <- if(metadata_position == "top") gg_metadata else gg_tmb
-  gg_bottom_plot <- if(metadata_position == "top") gg_tmb else gg_metadata
+  gg_top_height <- if(metadata_on_top) gg_metadata_height else gg_tmb_height
+  gg_bottom_height <- if(metadata_on_top) gg_tmb_height else gg_metadata_height
+  gg_top_plot <- if(metadata_on_top) gg_metadata else gg_tmb
+  gg_bottom_plot <- if(metadata_on_top) gg_tmb else gg_metadata
+
+  # Buffers between main plot and top & bottom margin plots (will shrink the main oncoplot)
+  buffer_top = if(metadata_on_top) buffer_metadata else buffer_tmb
+  buffer_bottom = if(metadata_on_top) buffer_tmb else buffer_metadata
+
+  buffer_top = if(is.null(gg_top_plot)) 0 else buffer_top
+  buffer_bottom = if(is.null(gg_bottom_plot)) 0 else buffer_bottom
 
   gg_main_height = 100 - gg_tmb_height - gg_metadata_height
-  gg_main_top = gg_top_height + 1
-  gg_main_bottom = gg_top_height + 1 + gg_main_height
+  gg_main_top = gg_top_height + buffer_top
+  gg_main_bottom = gg_main_top + gg_main_height - buffer_bottom
 
   gg_main_width = 100 - gg_gene_width
 
[…]
