+++ man/ggoncoplot_options.Rd
@@ -11,6 +11,8 @@ ggoncoplot_options(
   plotsize_tmb_rel_height = 10,
   plotsize_gene_rel_width = 20,
   plotsize_metadata_rel_height = 20,
+  buffer_metadata = 2,
+  buffer_tmb = 1,
   xlab_title = "Sample",
   ylab_title = "Gene",
   sample_id_position = c("bottom", "top"),
@@ -79,6 +81,8 @@ ggoncoplot_options(
 \item{plotsize_gene_rel_width}{percentage of horizontal space the gene barplot should take up. Must be some value between 5-90 (number)}
 
 \item{plotsize_metadata_rel_height}{percentage of vertical space the metadata tile plot should take up. Must be some value between 5-90 (number)}
+
+\item{buffer_metadata, buffer_tmb}{amount of space to add between the main oncoplot and tmb/metadata marginal plots (number)}
 
 \item{xlab_title}{x axis label. Set \code{xlab_title = NULL} to remove title (string)}
 
