+++ man/combine_plots.Rd
@@ -12,7 +12,9 @@ combine_plots(
   gg_tmb_height,
   gg_gene_width,
   gg_metadata_height,
-  metadata_position
+  metadata_position,
+  buffer_metadata,
+  buffer_tmb
 )
 }
 \arguments{
@@ -31,6 +33,8 @@ combine_plots(
 \item{gg_metadata_height}{percentage of plot height taken up by metadata plot (should be between 5-95) (number)}
 
 \item{metadata_position}{should metadata plot be on the 'top' or the 'bottom' of the oncoplot?}
+
+\item{buffer_metadata, buffer_tmb}{amount of space to add between the main oncoplot and tmb/metadata marginal plots (number)}
 }
 \value{
 patchwork object (or ggplot obj if both \code{gg_tmb} and \code{gg_gene} are NULL)
