### 2025-03-13 bfe365d
fix: metadata_position roxygen comment
