### 2025-09-14 #7390 @Kevin-Mattheus-Moerman
@luciorq,  @johnvusich, @vinegang, @Lesly1031, @mattwarkentin, @kelly-sovacool, @welch16, @chrisamiller  would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled _ggoncoplot: an R package for interactive visualisation of somatic mutation data from cancer patient cohorts_.

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/selkamand/ggoncoplot) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07390/joss.07390/10.21105.joss.07390.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

If you are interested you can let me know by responding here.

Note that this is an ongoing review where 1 reviewer is already finished but 1 other has not been responding. We therefore need at least 1 other reviewer who can hopefully complete the review in a timely manor.

Thanks!
