### 2025-09-27 #7390 @selkamand
Thanks @welch16 for the great suggestions. I've addressed them in https://github.com/selkamand/ggoncoplot/issues/203#issuecomment-3341629445 and in pull request https://github.com/selkamand/ggoncoplot/pull/204. Let me know what you think!
