### 2025-10-04 #7390 @vinegang
@selkamand

Apart from a small installation step (missing ggiraph and ggEDA in my env), everything else worked exactly as described. Once those dependencies were installed, it was smooth sailing — the manual was detailed, easy to follow, and it was genuinely a breeze to run the example steps and reproduce the plots. The customization options were also very well documented, making it simple to tweak and adapt the plots for different datasets.

Overall, the paper is concise, polished, and well written. The references look good, agree with @majensen suggestion to add  cBioPortal paper (Cerami et al., 2012). Great work — I really enjoyed reviewing this package!
